+++ test/server/fixtures/class-methods-project/lib/controllers/composeController.js
@@ -1,0 +1,15 @@
+export class ComposeController {
+    constructor() {
+        this.fromChannel = new ReactiveVar(null);
+    }
+
+    canSendEmail() {
+        return !!this.fromChannel.get();
+    }
+
+    hasSelectedUsers = () => false;
+
+    get selectedCount() {
+        return 0;
+    }
+}